third_party/blink/web_tests/http/tests/history/popstate-fires-with-pending-requests-expected.txt

Tests that popstate events fire when going back, even when there are resource requests pending.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting slow image request
Pushing newState1
Current search: ?newState1
Pushing newState2
Current search: ?newState2
Going back
PASS event.state is "newState1"
PASS successfullyParsed is true

TEST COMPLETE
